integrated with new API call
